Refactor: Split monolithic api.ts into domain files and standardize API error handling with ApiError
